requirements.txt
aiohttp==3.8.4
aioresponses==0.7.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==39.0.2
dnspython==2.3.0
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
httpretty==1.1.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==1.0.0
python-nmap==0.7.1
python-Wappalyzer==0.3.1
python-whois==0.8.0
requests==2.28.2
soupsieve==2.4
Sublist3r==1.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
